requirements.txt
MySQL-python==1.2.5
Flask==0.12
itsdangerous==0.24
Jinja2==2.9.5 
MarkupSafe==0.23 
Werkzeug==0.11.15 
click==6.7
cffi==1.9.1 
cryptography==1.7.2 
enum34==1.1.6 
idna==2.2 
ipaddress==1.0.18 
pyasn1==0.1.9 
pycparser==2.17 
pyopenssl==16.2.0
Flask==0.12
itsdangerous==0.24
Jinja2==2.9.5 
MarkupSafe==0.23 
Werkzeug==0.11.15 
click==6.7
dpkt==1.8.8
pycrypto==2.6.1
rsa==3.4.2
